bot/requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
google-api-python-client
google-auth
dateparser
pillow
httpx>=0.24.0
apscheduler>=3.10.0
